chore: move tests out of __tests__ like other apps
