+++ src/components/Form/Radio/index.test.js
@@ -4,7 +4,7 @@ import { MuiThemeProvider } from 'material-ui'
 import lightBaseTheme from 'material-ui/styles/baseThemes/lightBaseTheme'
 import getMuiTheme from 'material-ui/styles/getMuiTheme'
 import i18n from '@dhis2/d2-i18n/index'
-import Radio from '../'
+import Radio from './'
 
 const muiTheme = getMuiTheme(lightBaseTheme)
