+++ src/components/Form/Schemas/index.test.js
@@ -1,12 +1,12 @@
 import React from 'react'
 import { mount } from 'enzyme'
 import { Provider } from 'react-redux'
-import Schemas from '../'
-import { Loading } from '../../../Loading'
 import { MuiThemeProvider } from 'material-ui'
 import lightBaseTheme from 'material-ui/styles/baseThemes/lightBaseTheme'
 import getMuiTheme from 'material-ui/styles/getMuiTheme'
-import { store } from '../../../../store'
+import { Loading } from '../../Loading'
+import { store } from '../../../store'
+import Schemas from './'
 
 const muiTheme = getMuiTheme(lightBaseTheme)
