+++ src/components/Loading/index.test.js
@@ -3,7 +3,7 @@ import { mount } from 'enzyme'
 import { MuiThemeProvider } from 'material-ui'
 import lightBaseTheme from 'material-ui/styles/baseThemes/lightBaseTheme'
 import getMuiTheme from 'material-ui/styles/getMuiTheme'
-import { Loading } from '../'
+import { Loading } from './'
 
 const muiTheme = getMuiTheme(lightBaseTheme)
 let wrapper
